requirements.txt
openai>=1.43.0
azure-storage-blob>=12.20.0
SQLAlchemy>=2.0.30
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
pydantic>=2.7.0
tenacity>=8.5.0
PyYAML>=6.0.1
